Switch to MapboxGL. Need to add a key for this to work. Known not to work with point type maps.
For the most part, I left the old code in-place and commented it out, to (hopefully) make it easier to see diffs
A fair amount of code is now not used around map projections, since turf.js is doing that work.
However, I left the old code in-place, even though its calculations aren't used.